AP-23041: Fix docs of parse_int
AP-23041 (Expression: summary of parse_int says it creates float)

+++ org.knime.core.expressions/src/main/java/org/knime/core/expressions/functions/StringFunctions.java
@@ -1996,7 +1996,7 @@ public final class StringFunctions {
                 Convert a string to an integer if possible.
 
                 This function attempts to parse the input string as an integer. If
-                successful, returns the float representation of the string. If the
+                successful, returns the integer representation of the string. If the
                 input string cannot be parsed as an integer (or the input is
                 `MISSING`), returns `MISSING`.
